src/test/pyspark/tests/requirements.txt
pyspark
pytest
python-dotenv